requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.3.2
openpyxl==3.1.5
selenium==4.15.2
webdriver-manager==4.0.1
